move n returns to utils
